backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pymongo==4.6.0
motor==3.3.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
firebase-admin==6.4.0
APScheduler==3.10.4
requests==2.31.0
email-validator==2.1.0
python-dateutil==2.8.2 
aiohttp 